+++ atlasdb-cassandra/src/main/java/com/palantir/atlasdb/keyvalue/cassandra/CqlKeyValueService.java
@@ -33,7 +33,6 @@ import java.util.stream.Collectors;
 
 import javax.net.ssl.SSLContext;
 
-import org.apache.commons.lang3.Validate;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -281,13 +280,7 @@ public class CqlKeyValueService extends AbstractKeyValueService {
         Metadata metadata = cluster.getMetadata();
 
         final CassandraKeyValueServiceConfig config = configManager.getConfig();
-        String partitioner = metadata.getPartitioner();
-        if (!config.safetyDisabled()) {
-            Validate.isTrue(
-                    CassandraConstants.ALLOWED_PARTITIONERS.contains(partitioner),
-                    "partitioner is: " + partitioner);
-        }
-
+        CassandraVerifier.validatePartitioner(metadata.getPartitioner(), config);
 
         Set<Peer> peers = CqlKeyValueServices.getPeers(session);