+++ atlasdb-container-test-utils/src/main/java/com/palantir/atlasdb/containers/ThreeNodeCassandraCluster.java
@@ -57,7 +57,6 @@ public class ThreeNodeCassandraCluster extends Container {
             .mutationBatchCount(10000)
             .mutationBatchSizeBytes(10000000)
             .fetchBatchCount(1000)
-            .safetyDisabled(false)
             .autoRefreshNodes(false)
             .build();